Finish cert import and add mocks and unit tests

+++ pkg/kyma/cmd/install/cmd.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"io"
 	"io/ioutil"
-	"log"
 	"net/http"
 	"os"
 	"path/filepath"
@@ -32,7 +31,6 @@ import (
 
 type command struct {
 	opts *Options
-	cert trust.Certifier
 	core.Command
 }
 
@@ -68,7 +66,6 @@ func NewCmd(o *Options) *cobra.Command {
 
 	cmd := command{
 		Command: core.Command{Options: o.Options},
-		cert:    trust.NewCertifier(),
 		opts:    o,
 	}
 
@@ -154,15 +151,10 @@ func (cmd *command) Run() error {
 		if err := cmd.waitForInstaller(); err != nil {
 			return err
 		}
-
-		s = cmd.NewStep("Importing Kyma root certificate")
-		if err := cmd.importCertificate(); err != nil {
-			cmd.CurrentStep.Failure()
-			return err
-		}
-		s.Successf("Kyma root certificate imported")
-	} else {
-		cmd.cert.Instructions()
+	}
+	if err := cmd.importCertificate(trust.NewCertifier(cmd.opts.Verbose)); err != nil {
+		// certificate import errors do not mean installation failed
+		cmd.CurrentStep.LogError(err.Error())
 	}
 
 	if err := cmd.printSummary(); err != nil {
@@ -913,32 +905,35 @@ func (cmd *command) patchMinikubeIP() error {
 	return nil
 }
 
-func (cmd *command) importCertificate() error {
-	// get cert from cluster
-	cert, err := cmd.Kubectl().RunCmd("get", "configmap", "net-global-overrides", "-n", "kyma-installer", "-o", "jsonpath='{.data.global\\.ingress\\.tlsCrt}'")
-	if err != nil {
-		return err
+func (cmd *command) importCertificate(ca trust.Certifier) error {
+	if !cmd.opts.NoWait {
+		// get cert from cluster
+		cert, err := ca.Certificate()
+		if err != nil {
+			return err
+		}
+
+		tmpFile, err := ioutil.TempFile(os.TempDir(), "kyma-*.crt")
+		if err != nil {
+			return errors.Wrap(err, "Cannot create temporary file for Kyma certificate")
+		}
+		defer os.Remove(tmpFile.Name())
+
+		if _, err = tmpFile.Write(cert); err != nil {
+			return errors.Wrap(err, "Failed to write the kyma certificate")
+		}
+		if err := tmpFile.Close(); err != nil {
+			return err
+		}
+
+		if err := ca.StoreCertificate(tmpFile.Name(), cmd.CurrentStep); err != nil {
+			return err
+		}
+		cmd.CurrentStep.Successf("Kyma root certificate imported")
+
+	} else {
+		cmd.CurrentStep.LogError(ca.Instructions())
 	}
 
-	decodedCert, err := base64.StdEncoding.DecodeString(cert)
-	if err != nil {
-		return err
-	}
-	tmpFile, err := ioutil.TempFile(os.TempDir(), "kyma-*.crt")
-	if err != nil {
-		log.Fatal("Cannot create temporary file for Kyma certificate", err)
-	}
-	defer os.Remove(tmpFile.Name())
-
-	if _, err = tmpFile.Write(decodedCert); err != nil {
-		log.Fatal("Failed to write the kyma certificate", err)
-	}
-	if err := tmpFile.Close(); err != nil {
-		return err
-	}
-
-	if err := cmd.cert.StoreCertificate(tmpFile.Name()); err != nil {
-		return err
-	}
 	return nil
 }
